Move tests for validation in EntityAliasHandler to separate class
